requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
yfinance==0.2.54
python-dotenv==1.0.1
